Split ctf/views.py into ctf/views package (SonarCloud S104)
Decompose the ~4,400-line ctf/views.py into a ctf/views/ package of
cohesive, sub-500-line submodules, clearing python:S104. Behavior-
preserving: the URLconf still resolves views.<name> via the package
facade, which re-exports only the public URLconf views (no trampolines,
no framework objects or private helpers on the facade). Shared access /
parsing helpers live in _access.py / _parsing.py; the JSON API moves to
an api/ subpackage.

Per the #911 quality bar, test patch targets that pointed at first-party
ctf.views.* topology were removed rather than relocated: organizer-role
tests now drive the real get_user_role via group membership, and the
render / participant / bracket paths run for real under
@pytest.mark.django_db with the conftest fixtures (services mocked only
at their already-tracked boundary). The 11 corresponding ctf.views.*
entries are dropped from the ADR-019 boundary-mock baseline.
