+++ shifter/shifter_platform/ctf/views/api/__init__.py
@@ -1,0 +1,1 @@
+"""CTF JSON API views (split from ctf/views.py, issue #885)."""
