+++ shifter/shifter_platform/tests/ctf/test_event_deletion.py
@@ -34,6 +34,8 @@ from ctf.enums import EventStatus
 @pytest.fixture
 def mock_user():
     """Create a mock authenticated user (organizer)."""
+    from shared.auth import CTF_ORGANIZER_GROUP
+
     user = MagicMock()
     user.pk = 1
     user.id = 1
@@ -45,6 +47,9 @@ def mock_user():
     user.is_staff = False
     user.is_superuser = False
     user.backend = "django.contrib.auth.backends.ModelBackend"
+    # Drive the real get_user_role via group membership (the public
+    # get_user_group_names contract) instead of patching first-party topology.
+    user.groups.values_list.return_value = [CTF_ORGANIZER_GROUP]
     return user
 
 
@@ -62,6 +67,8 @@ def mock_standard_user():
     user.is_staff = False
     user.is_superuser = False
     user.backend = "django.contrib.auth.backends.ModelBackend"
+    # No CTF groups -> real get_user_role reports neither organizer nor participant.
+    user.groups.values_list.return_value = []
     return user
 
 
@@ -151,10 +158,6 @@ def _mock_auth_organizer(mock_user):
 
     Also patches context processors that would otherwise hit the DB.
     """
-    from ctf.bridges import UserRole
-
-    role = UserRole(is_ctf_organizer=True, is_ctf_participant=False, active_ctf_event=None)
-
     ctx_proc_defaults = {
         "is_ctf_user": True,
         "is_ctf_organizer": True,
@@ -170,7 +173,6 @@ def _mock_auth_organizer(mock_user):
     }
 
     with (
-        patch("ctf.views.get_user_role", return_value=role),
         patch("django.contrib.auth.get_user", return_value=mock_user),
         patch("django.contrib.auth.middleware.get_user", return_value=mock_user),
         patch("ctf.context_processors.ctf_navigation", return_value=ctx_proc_defaults),
@@ -183,12 +185,7 @@ def _mock_auth_organizer(mock_user):
 @pytest.fixture
 def _mock_auth_standard(mock_standard_user):
     """Patch Django auth to authenticate mock_standard_user as non-organizer."""
-    from ctf.bridges import UserRole
-
-    role = UserRole(is_ctf_organizer=False, is_ctf_participant=False, active_ctf_event=None)
-
     with (
-        patch("ctf.views.get_user_role", return_value=role),
         patch("django.contrib.auth.get_user", return_value=mock_standard_user),
         patch("django.contrib.auth.middleware.get_user", return_value=mock_standard_user),
     ):
